requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
greenlet==2.0.2
idna==3.4
psycopg2-binary==2.9.6
pyTelegramBotAPI==4.11.0
requests==2.28.2
SQLAlchemy==2.0.10
telebot-calendar==1.2
typing_extensions==4.5.0
urllib3==1.26.15
